cifar10-denoise and mnist-denoise: switch to instance mode
Running average mode doesn't seem to produce the expected
result for the mnist-denoise Yg output after switching to
Adam optimizer. The issue only seems to occur when importing
models (mnist-disc) while models created from scratch seem
to be working fine (e.g. mnist-denoise). To see the issue,
compare the Yg output with running averages vs instance
normalization. More investigation is needed to determine if
this is actually a bug or if this is a result of the well
known side effect of Batch Normalization producing different
results between testing and inference.